+++ tests/test_atari.py
@@ -14,7 +14,7 @@ import stable_baselines.trpo_mpi.run_atari as trpo_atari
 
 ENV_ID = 'BreakoutNoFrameskip-v4'
 SEED = 3
-NUM_TIMESTEPS = 500
+NUM_TIMESTEPS = 300
 NUM_CPU = 2
 
 
@@ -62,7 +62,7 @@ def test_deepq():
     env = wrap_atari_dqn(env)
 
     model = DQN(env=env, policy=CnnPolicy, learning_rate=1e-4, buffer_size=10000, exploration_fraction=0.1,
-                exploration_final_eps=0.01, train_freq=4, learning_starts=10000, target_network_update_freq=1000,
+                exploration_final_eps=0.01, train_freq=4, learning_starts=100, target_network_update_freq=100,
                 gamma=0.99, prioritized_replay=True, prioritized_replay_alpha=0.6)
     model.learn(total_timesteps=NUM_TIMESTEPS)