+++ tests/test_save.py
@@ -12,7 +12,7 @@ from stable_baselines.common.vec_env import DummyVecEnv
 from stable_baselines.common.evaluation import evaluate_policy
 from stable_baselines.common.policies import MlpPolicy, FeedForwardPolicy
 
-N_EVAL_EPISODES = 100
+N_EVAL_EPISODES = 5
 
 MODEL_LIST = [
     A2C,
@@ -54,12 +54,29 @@ def test_model_manipulation(request, model_class, storage_method, store_format):
     model_fname = './test_model_{}.model'.format(request.node.name)
     store_as_cloudpickle = store_format == "cloudpickle"
 
+    kwargs = dict(seed=0, gamma=0.4)
+    if model_class in [DQN]:
+        kwargs["learning_starts"] = 0
+        kwargs["exploration_final_eps"] = 0.05
+
+    if model_class == PPO1:
+        kwargs["entcoeff"] = 0.0
+        kwargs["optim_batchsize"] = 4
+        kwargs["timesteps_per_actorbatch"] = 4
+
+    if model_class in [A2C, ACKTR, PPO2]:
+        kwargs["n_steps"] = 4
+        kwargs["ent_coef"] = 0.0
+
+    if model_class in [TRPO]:
+        kwargs["timesteps_per_batch"] = 4
+
     try:
         env = DummyVecEnv([lambda: IdentityEnv(10)])
 
         # create and train
-        model = model_class(policy="MlpPolicy", env=env, seed=0)
-        model.learn(total_timesteps=10000)
+        model = model_class(policy="MlpPolicy", env=env, **kwargs)
+        model.learn(total_timesteps=15)
 
         env.envs[0].action_space.seed(0)
         mean_reward, _ = evaluate_policy(model, env, deterministic=True,
@@ -68,8 +85,10 @@ def test_model_manipulation(request, model_class, storage_method, store_format):
         # test action probability for given (obs, action) pair
         env = model.get_env()
         obs = env.reset()
-        observations = np.array([obs for _ in range(10)])
+        observations = np.array([env.step([env.action_space.sample()])[0] for _ in range(10)])
         observations = np.squeeze(observations)
+        selected_actions, _ = model.predict(observations, deterministic=True)
+
         actions = np.array([env.action_space.sample() for _ in range(10)])
         actions_probas = model.action_probability(observations, actions=actions)
         assert actions_probas.shape == (len(actions), 1), actions_probas.shape
@@ -99,22 +118,12 @@ def test_model_manipulation(request, model_class, storage_method, store_format):
         env = DummyVecEnv([lambda: IdentityEnv(10)])
         model.set_env(env)
 
-        # predict the same output before saving
-        env.envs[0].action_space.seed(0)
-        loaded_mean_reward, _ = evaluate_policy(model, env, deterministic=True, n_eval_episodes=N_EVAL_EPISODES)
-        # Allow 10% diff
-        assert abs((mean_reward - loaded_mean_reward) / mean_reward) < 0.1, "Error: the prediction seems to have changed between " \
-                                                                            "loading and saving"
+        # check if model still selects the same actions
+        new_selected_actions, _ = model.predict(observations, deterministic=True)
+        assert np.allclose(selected_actions, new_selected_actions, 1e-4)
 
         # learn post loading
-        model.learn(total_timesteps=100)
-
-        # validate no reset post learning
-        env.envs[0].action_space.seed(0)
-        loaded_mean_reward, _ = evaluate_policy(model, env, deterministic=True, n_eval_episodes=N_EVAL_EPISODES)
-
-        assert abs((mean_reward - loaded_mean_reward) / mean_reward) < 0.15, "Error: the prediction seems to have changed between " \
-                                                                            "pre learning and post learning"
+        model.learn(total_timesteps=15)
 
         # predict new values
         evaluate_policy(model, env, n_eval_episodes=N_EVAL_EPISODES)