+++ docs/misc/changelog.rst
@@ -93,6 +93,7 @@ Documentation:
 - Added Slime Volleyball project (@hardmaru)
 - Added a table of the variables accessible from the ``on_step`` function of the callbacks for each algorithm (@PartiallyTyped)
 - Fix typo in README.md (@ColinLeongUDRI)
+- Fix typo in gail.rst (@roccivic)
 
 Release 2.10.0 (2020-03-11)
 ---------------------------
@@ -749,3 +750,4 @@ Thanks to @bjmuld @iambenzo @iandanforth @r7vme @brendenpetersen @huvar @abhiskk
 @MarvineGothic @jdossgollin @SyllogismRXS @rusu24edward @jbulow @Antymon @seheevic @justinkterry @edbeeching
 @flodorner @KuKuXia @NeoExtended @PartiallyTyped @mmcenta @richardwu @tirafesi @caburu @johannes-dornheim @kvenkman @aakash94
 @enderdead @hardmaru @jbarsce @ColinLeongUDRI @shwang @YangRui2015 @sophiagu @OGordon100 @SVJayanthi @sunshineclt
+@roccivic