+++ docs/misc/changelog.rst
@@ -28,6 +28,7 @@ Deprecations:
 
 Others:
 ^^^^^^^
+- Faster tests, switched to GitHub CI
 
 Documentation:
 ^^^^^^^^^^^^^^